requirements.txt
Flask==2.3.3
moviepy==1.0.3
google-genai
opencv-python==4.8.1.78
Pillow==10.0.1
numpy==1.24.3
requests==2.31.0
python-dotenv==1.0.0
wave
